requirement.txt
python-dotenv
requests
beautifulsoup4
black
openai